refactor: ascend `RecordCloneFlag` into parent level
XRECORD also uses this constant
